requirements-hosaka.txt
fastapi>=0.116,<1.0
uvicorn>=0.35,<1.0
pydantic>=2.11,<3.0
python-multipart>=0.0.9,<1.0
httpx>=0.27,<1.0
python-dotenv>=1.0,<2.0
